move plot module
