+++ autofit/non_linear/plot/__init__.py
@@ -1,0 +1,2 @@
+from autofit.non_linear.plot.samples_plotters import SamplesPlotter
+from autofit.non_linear.plot.output import Output
