+++ autofit/non_linear/search/mcmc/emcee/plotter.py
@@ -1,7 +1,7 @@
 import numpy as np
 import corner
 
-from autofit.plot.samples_plotters import MCMCPlotter
+from autofit.non_linear.plot.samples_plotters import MCMCPlotter
 
 class EmceePlotter(MCMCPlotter):
