+++ autofit/non_linear/search/mcmc/emcee/search.py
@@ -16,7 +16,7 @@ from autofit.non_linear.search.mcmc.auto_correlations import AutoCorrelations
 from autofit.non_linear.samples.sample import Sample
 from autofit.non_linear.samples.mcmc import SamplesMCMC
 from autofit.plot import EmceePlotter
-from autofit.plot.output import Output
+from autofit.non_linear.plot.output import Output
 
 
 class Emcee(AbstractMCMC):
