+++ autofit/non_linear/search/mcmc/zeus/plotter.py
@@ -1,4 +1,4 @@
-from autofit.plot.samples_plotters import MCMCPlotter
+from autofit.non_linear.plot.samples_plotters import MCMCPlotter
 
 
 class ZeusPlotter(MCMCPlotter):
