+++ autofit/non_linear/search/nest/nautilus/search.py
@@ -14,7 +14,7 @@ from autofit.non_linear.paths.null import NullPaths
 from autofit.non_linear.search.nest import abstract_nest
 from autofit.non_linear.samples.sample import Sample
 from autofit.non_linear.samples.nest import SamplesNest
-from autofit.plot.output import Output
+from autofit.non_linear.plot.output import Output
 
 logger = logging.getLogger(__name__)
