+++ autofit/non_linear/search/nest/ultranest/plotter.py
@@ -1,7 +1,7 @@
 from functools import wraps
 from autofit.plot import SamplesPlotter
 
-from autofit.plot.samples_plotters import skip_plot_in_test_mode
+from autofit.non_linear.plot.samples_plotters import skip_plot_in_test_mode
 
 def log_value_error(func):
