+++ autofit/non_linear/search/nest/ultranest/search.py
@@ -10,7 +10,7 @@ from autofit.non_linear.fitness import Fitness
 from autofit.non_linear.samples.sample import Sample
 from autofit.non_linear.samples.nest import SamplesNest
 from autofit.plot import UltraNestPlotter
-from autofit.plot.output import Output
+from autofit.non_linear.plot.output import Output
 
 class UltraNest(abstract_nest.AbstractNest):
     __identifier_fields__ = (
