+++ autofit/non_linear/search/optimize/drawer/plotter.py
@@ -1,4 +1,4 @@
-from autofit.plot.samples_plotters import MCMCPlotter
+from autofit.non_linear.plot.samples_plotters import MCMCPlotter
 
 class DrawerPlotter(MCMCPlotter):
