--- autofit/plot/__init__.py
@@ -1,2 +1,0 @@
-from autofit.plot.samples_plotters import SamplesPlotter
-from autofit.plot.output import Output
